feat(utopia-scalar): allow overriding default html for scalar.
At one point, there was a bug in the published verison of scalar and,
since the default html references the latest version from the cdn, it
caused an outage for users despite no local code changes.

This change allows for specifying your own html, which, in addition to
providing an avenue for pinning the scalar version, allows for other
customizations.